profiles.desc: demote all profiles to exp
I don't remember the reason to initially use "dev" for all the profiles,
maybe because of repoman or eselect profile not allowing me to select
exp profiles? There really is no reason for them not to be exp, though,
so fix the longtime "mistake".

This also concludes the last loong profile work before pushing to
upstream.

Closes: #19
Signed-off-by: WANG Xuerui <xen0n@gentoo.org>